requirements.txt
beautifulsoup4>=4.9.3
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
jinja2>=3.0.0
google-generativeai>=0.3.0
pyarrow>=8.0.0
tqdm>=4.62.0
requests>=2.26.0
python-dotenv>=0.19.0
colorama>=0.4.4
sqlalchemy>=2.0.0
alembic>=1.11.0
pytest>=6.2.5
flask>=2.0.0
flask-cors>=3.0.10
python-dateutil>=2.8.2
ipython>=8.0.0
huggingface-hub>=0.13.0
typer>=0.9.0 